refactor: unificación de Perfil en un único componente (se eliminan subcomponentes separados)
