--- frontend/src/components/app/Perfil/avatars.ts
@@ -1,8 +1,0 @@
-
-export const avatars = [
-  "/avatars/avatar1.png",
-  "/avatars/avatar2.png",
-  "/avatars/avatar3.png",
-  "/avatars/avatar4.png",
-  "/avatars/avatar5.png",
-];
